feat: add reactive repository to hold TimerEntity data
BREAKING CHANGES: The `TimerCounterBloc` and `TimerBloc` has been
updated to use `TimerStorageRepository` to set/get data from. Previously the `TimerCounterBloc` and `TimerBloc` used to hold the data from it's own classes and passing from constructor/event. This was not a recommended way to do those sort of things as the BLoC website said itself, and it introduce a bug into the application.